Added the commute time computations and the main function
The cars class was changed to include a spawn variable and a helper function to retrieve the value of spawn. The Poisson class was changed to accept the time as a parameter and provide it to the cars it spawns. The Gameboard class was changed so that get_rid calculates the commute of the car that is being removed. time_seg was also updated to return an array of the commute times for the cars that exited during that cycle. Main.py was made to run the simulation and print the average commute time. Time is in seconds.